testinfra/requirements.txt
# Direct Dependencies
testinfra==1.8.0
pytest==3.2.2
paramiko==2.3.3
pyyaml==3.13

# Transitive Dependencies
cffi==1.11.0
asn1crypto==0.23.0
bcrypt==3.1.3
cryptography==2.3
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
py==1.4.34
pyasn1==0.3.6
pycparser==2.18
PyNaCl==1.1.2
six==1.11.0
